+++ packages/shared/trophies/getTrophiesCategory.ts
@@ -1,4 +1,4 @@
-import { ARAM, SUMMONERS_RIFT, TWISTED_TREELINE } from './types';
+import { ARAM, SUMMONERS_RIFT } from './types';
 
 import { MATCHMAKING } from '../riot-api/gameConstants';
 
@@ -11,7 +11,6 @@ const summonersRift = [
 ];
 
 const aram = [MATCHMAKING.ARAM_5v5, MATCHMAKING.BILGEWATER_ARAM_5x5];
-const twistedTreeline = [MATCHMAKING.BLIND_PICK_3v3, MATCHMAKING.RANKED_3v3];
 
 const getTrophiesCategory = queueId => {
   if (summonersRift.includes(queueId)) {
@@ -19,9 +18,6 @@ const getTrophiesCategory = queueId => {
   }
   if (aram.includes(queueId)) {
     return ARAM;
-  }
-  if (twistedTreeline.includes(queueId)) {
-    return TWISTED_TREELINE;
   }
 };